backend/pyproject.toml
[project]
name = "supply-chain-backend"
version = "1.0.0"
description = "Backend API for supply chain optimization with agent workflows"
authors = [
    {name = "Supply Chain Team"}
]
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "pydantic>=2.6.0",
    "PyJWT[crypto]==2.10.1",
    "passlib[bcrypt]==1.7.4",
    "python-multipart==0.0.6",
    "websockets==12.0",
    "python-dotenv==1.0.0",
    "requests==2.32.5",
    "a2a-sdk>=0.3.0",
    "httpx>=0.28.1",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "opentelemetry-instrumentation-httpx>=0.40b0",
    "opentelemetry-instrumentation-fastapi>=0.40b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
requires-python = ">=3.12"

[project.scripts]
start = "app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
